+++ frontend/src/App.tsx
@@ -10,10 +10,10 @@ import BattleResultPage from './pages/battleResultPage';
 import LoginPage from './pages/loginPage';
 import NicknamePage from './pages/nicknamePage';
 import EntryPage from './pages/entryPage';
-import TutorialTeamSelectPage from './pages/tutorialTeamSelectPage';
+import TutorialTeamSelectPage from './pages/tutorialPage/teamSelect';
 import ErrorPage from './pages/errorPage';
-import TutorialBattlePage from './pages/tutorialBattlePage';
-import { TUTORIAL_BATTLE_INFO } from './pages/tutorial/const/tutorialBattle';
+import TutorialBattlePage from './pages/tutorialPage/battle';
+import { TUTORIAL_BATTLE_INFO } from './pages/tutorialPage/const/tutorialBattle';
 import { ToastContainer } from './commons/components/toast/ToastContainer';
 import { useAuthStore } from './commons/stores/authStore';
 import './App.css';
