+++ frontend/src/pages/tutorialPage/battle/components/PracticeGuideCard.tsx
