+++ frontend/src/pages/tutorialPage/battle/hooks/useAutoExitOnDone.ts
