+++ frontend/src/pages/tutorialPage/battle/hooks/usePracticeFlow.ts
@@ -3,7 +3,7 @@ import type { BattlePhase } from '@/commons/types/battle';
 import { useBattleStore, selectBattleProgress } from '@/pages/battlePage/stores/battleStore';
 import { useAuthStore } from '@/commons/stores/authStore';
 import { soundManager } from '@/commons/utils/soundManager';
-import { TUTORIAL_ATTACK_MESSAGES, TUTORIAL_BATTLE_ID } from '@/pages/tutorial/const/tutorialBattle';
+import { TUTORIAL_ATTACK_MESSAGES, TUTORIAL_BATTLE_ID } from '@/pages/tutorialPage/const/tutorialBattle';
 
 const ATTACK_DISCUSSIONS = [
   {
