+++ frontend/src/pages/tutorialPage/battle/hooks/useTutorialUI.ts
