+++ frontend/src/pages/tutorialPage/const/tutorialBattle.ts
